refactor(yarn-plugin-renderer): separate Next artifact owners
