+++ yarn/plugin-renderer/sources/artifact/source.interfaces.ts
@@ -1,0 +1,10 @@
+import type { PortablePath } from '@yarnpkg/fslib'
+
+export interface RendererArtifactSource {
+  readonly appCwd: PortablePath
+  readonly appRelativeCwd: PortablePath
+  readonly nextOutputCwd: PortablePath
+  readonly nextOutputRelativeCwd: PortablePath
+  readonly standaloneAppCwd: PortablePath
+  readonly standaloneCwd: PortablePath
+}
