--- yarn/plugin-renderer/sources/commands/renderer-dev.command.test.ts
@@ -1,12 +1,0 @@
-import assert                    from 'node:assert/strict'
-import { test }                  from 'node:test'
-
-import { createRendererDevArgs } from './renderer-dev.command.js'
-
-test('should run Next dev from renderer source app directory', () => {
-  assert.deepEqual(createRendererDevArgs(undefined), ['next', 'dev', 'src'])
-})
-
-test('should use explicit webpack dev arguments for Next versions after 15', () => {
-  assert.deepEqual(createRendererDevArgs('16.2.10'), ['next', 'dev', 'src', '--webpack'])
-})
