+++ yarn/plugin-renderer/sources/integrations/next/execution/environment.interfaces.ts
@@ -1,0 +1,11 @@
+import type { PortablePath } from '@yarnpkg/fslib'
+
+export interface NextExecutionEnvironmentOptions {
+  readonly nextConfigAdapterPath?: PortablePath
+  readonly output?: string
+}
+
+export interface ExtractedNodeLoader {
+  readonly loader: string | undefined
+  readonly nodeOptions: string | undefined
+}
