+++ yarn/plugin-renderer/sources/integrations/next/execution/loader.constants.ts
@@ -7,4 +7,3 @@ export const NEXT_PACKAGE_PATH = '/node_modules/next/'
 export const NEXT_REQUIRE_CACHE_PATH = '/dist/server/dev/require-cache.js'
 export const NEXT_SERVER_CONFIG_PATH = '/dist/server/config.js'
 export const NEXT_WEBPACK_CONFIG_PATH = '/dist/build/webpack-config.js'
-export const RENDERER_STANDALONE_SERVER_ENTRYPOINT = 'index.cjs'
