+++ yarn/plugin-renderer/sources/integrations/next/standalone/discovery.interfaces.ts
@@ -1,0 +1,8 @@
+import type { PortablePath } from '@yarnpkg/fslib'
+
+export interface NextStandaloneManifestState {
+  readonly mtimeMs: number
+  readonly size: number
+}
+
+export type NextStandaloneManifestSnapshot = ReadonlyMap<PortablePath, NextStandaloneManifestState>
